refactor(task-graph): extract pipe/parallel/getLastTask/connect to WorkflowPipe.ts
The four module-level builder helpers are independent of the Workflow
class — the class instance and static methods already delegate to them.
Moving them to a sibling file trims Workflow.ts further as part of the
larger decomposition. Re-exported via common.ts so the public surface of
@workglow/task-graph is unchanged.
