+++ packages/task-graph/src/common.ts
@@ -23,6 +23,7 @@ export * from "./task-graph/Conversions";
 export * from "./task-graph/GraphToWorkflowCode";
 export * from "./task-graph/IWorkflow";
 export * from "./task-graph/Workflow";
+export * from "./task-graph/WorkflowPipe";
 
 export * from "./task-graph/TransformRegistry";
 export * from "./task-graph/TransformTypes";
