+++ packages/task-graph/src/task-graph/Workflow.ts
@@ -29,95 +29,12 @@ import {
   hasVectorOutput,
   updateBoundaryTaskSchemas,
 } from "./GraphSchemaUtils";
-import type { ITaskGraph } from "./ITaskGraph";
 import type { IWorkflow, WorkflowRunConfig } from "./IWorkflow";
 import { TaskGraph } from "./TaskGraph";
 import type { PropertyArrayGraphResult } from "./TaskGraphRunner";
 import { CompoundMergeStrategy, PROPERTY_ARRAY } from "./TaskGraphRunner";
 import type { ITransformStep } from "./TransformTypes";
-
-// ============================================================================
-// Standalone utility functions (moved from Conversions.ts to break circular
-// dependency — these need both Workflow and GraphAsTask which live here)
-// ============================================================================
-
-export function getLastTask(workflow: IWorkflow): ITask<any, any, any> | undefined {
-  const tasks = workflow.graph.getTasks();
-  return tasks.length > 0 ? tasks[tasks.length - 1] : undefined;
-}
-
-export function connect(
-  source: ITask<any, any, any>,
-  target: ITask<any, any, any>,
-  workflow: IWorkflow<any, any>
-): void {
-  workflow.graph.addDataflow(new Dataflow(source.id, "*", target.id, "*"));
-}
-
-export function pipe<A extends DataPorts, B extends DataPorts>(
-  [fn1]: [Taskish<A, B>],
-  workflow?: IWorkflow<A, B>
-): IWorkflow<A, B>;
-
-export function pipe<A extends DataPorts, B extends DataPorts, C extends DataPorts>(
-  [fn1, fn2]: [Taskish<A, B>, Taskish<B, C>],
-  workflow?: IWorkflow<A, C>
-): IWorkflow<A, C>;
-
-export function pipe<
-  A extends DataPorts,
-  B extends DataPorts,
-  C extends DataPorts,
-  D extends DataPorts,
->(
-  [fn1, fn2, fn3]: [Taskish<A, B>, Taskish<B, C>, Taskish<C, D>],
-  workflow?: IWorkflow<A, D>
-): IWorkflow<A, D>;
-
-export function pipe<
-  A extends DataPorts,
-  B extends DataPorts,
-  C extends DataPorts,
-  D extends DataPorts,
-  E extends DataPorts,
->(
-  [fn1, fn2, fn3, fn4]: [Taskish<A, B>, Taskish<B, C>, Taskish<C, D>, Taskish<D, E>],
-  workflow?: IWorkflow<A, E>
-): IWorkflow<A, E>;
-
-export function pipe<
-  A extends DataPorts,
-  B extends DataPorts,
-  C extends DataPorts,
-  D extends DataPorts,
-  E extends DataPorts,
-  F extends DataPorts,
->(
-  [fn1, fn2, fn3, fn4, fn5]: [
-    Taskish<A, B>,
-    Taskish<B, C>,
-    Taskish<C, D>,
-    Taskish<D, E>,
-    Taskish<E, F>,
-  ],
-  workflow?: IWorkflow<A, F>
-): IWorkflow<A, F>;
-
-export function pipe<I extends DataPorts, O extends DataPorts>(
-  args: Taskish<I, O>[],
-  workflow: IWorkflow<I, O> = new Workflow<I, O>()
-): IWorkflow<I, O> {
-  let previousTask = getLastTask(workflow);
-  const tasks = args.map((arg) => ensureTask(arg));
-  tasks.forEach((task) => {
-    workflow.graph.addTask(task);
-    if (previousTask) {
-      connect(previousTask, task, workflow);
-    }
-    previousTask = task;
-  });
-  return workflow;
-}
+import { getLastTask, parallel, pipe } from "./WorkflowPipe";
 
 /** Options accepted by {@link Workflow.rename}. */
 export interface RenameOptions {
@@ -125,29 +42,6 @@ export interface RenameOptions {
   readonly index?: number;
   /** Transform chain applied to the dataflow edge this rename creates. */
   readonly transforms?: ReadonlyArray<ITransformStep>;
-}
-
-export function parallel<I extends DataPorts = DataPorts, O extends DataPorts = DataPorts>(
-  args: (PipeFunction<I, O> | ITask<I, O> | IWorkflow<I, O> | ITaskGraph)[],
-  mergeFn: CompoundMergeStrategy = PROPERTY_ARRAY,
-  workflow: IWorkflow<I, O> = new Workflow<I, O>()
-): IWorkflow<I, O> {
-  let previousTask = getLastTask(workflow);
-  const tasks = args.map((arg) => ensureTask(arg));
-  const config = {
-    compoundMerge: mergeFn,
-  };
-  const name = `‖${args.map((_arg) => "𝑓").join("‖")}‖`;
-  class ParallelTask extends GraphAsTask<I, O> {
-    public static override type = name;
-  }
-  const mergeTask = new ParallelTask(config);
-  mergeTask.subGraph!.addTasks(tasks);
-  workflow.graph.addTask(mergeTask);
-  if (previousTask) {
-    connect(previousTask, mergeTask, workflow);
-  }
-  return workflow;
 }
 
 // Type definitions for the workflow
